move imu functions to robot classes for better reuse.
Also clean up some comments
