+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robot/constants.java
@@ -7,6 +7,10 @@ public class constants {
         leftFront, leftBack, rightFront, rightBack;
 
     }
+    public enum OrientationMode {
+        FIELD,
+        ROBOT
+    }
     public static final double ENCODER_TICKS = 537.70;
     public static final int TELEMETRY_MS_TRANSMISSION_INTERVAL = 25;
